backend/requirements.txt
Flask==2.0.1
requests
python-dotenv
